Add a script for running the test suites
