--- script/test
@@ -1,3 +1,0 @@
-#! /usr/bin/env bash
-
-python -m unittest discover -s ./test
