+++ test
@@ -1,0 +1,3 @@
+#! /usr/bin/env bash
+
+python -m unittest discover -s ./tests
