+++ tests/test_example.py
@@ -1,0 +1,7 @@
+import unittest
+
+from helpers.parameterized_tests import ParameterizedTest
+
+class TestTheTruth(ParameterizedTest):
+    def test_truth(self):
+            self.assertTrue(True)
